+++ backend/app/builders/ProcessoBuilder.java
@@ -782,6 +782,19 @@ public class ProcessoBuilder extends CriteriaBuilder<Processo> {
 
 	}
 
+	public ProcessoBuilder filtrarIdPresidente(Long idUsuarioPresidente) {
+
+		if (idUsuarioPresidente != null) {
+
+			addObjetoTramitavelAlias();
+			addRestriction(Restrictions.eq(OBJETO_TRAMITAVEL_ALIAS + ".usuarioResponsavel.id", idUsuarioPresidente));
+
+		}
+
+		return this;
+
+	}
+
 	public ProcessoBuilder filtrarPorPeriodoProcesso(Date periodoInicial, Date periodoFinal) {
 
 		if (periodoInicial != null) {
@@ -1137,6 +1150,7 @@ public class ProcessoBuilder extends CriteriaBuilder<Processo> {
 		public boolean isAnaliseGeoOpcional;
 		public boolean isGerente;
 		public boolean isDiretor;
+		public boolean isPresidente;
 		public Long idAnalistaGeo;
 		public String siglaSetorGerencia;
 		public String siglaSetorCoordenadoria;