+++ frontend/src/app.js
@@ -169,7 +169,7 @@ licenciamento.controller("AppController", ["$injector", "$scope", "$rootScope", 
 					app.utils.Perfis.ANALISTA_GEO,
 					app.utils.Perfis.ANALISTA_CAR,
 					app.utils.Perfis.DIRETOR,
-						app.utils.Perfis.ANALISTA_TECNICO
+					app.utils.Perfis.ANALISTA_TECNICO
 					].indexOf($rootScope.usuarioSessao.usuarioEntradaUnica.perfilSelecionado.codigo) > -1;
 			},
 			condicaoTramitacao: function() {
@@ -188,6 +188,7 @@ licenciamento.controller("AppController", ["$injector", "$scope", "$rootScope", 
 					return app.utils.CondicaoTramitacao.AGUARDANDO_ANALISE_GEO;
 				else if ($rootScope.usuarioSessao.usuarioEntradaUnica.perfilSelecionado.codigo === app.utils.Perfis.DIRETOR)
 					return app.utils.CondicaoTramitacao.AGUARDANDO_VALIDACAO_DIRETORIA;
+
 			},
 			deveFiltrarPorUsuario: true,
 			codigoPerfilSelecionado: function(){
@@ -195,6 +196,44 @@ licenciamento.controller("AppController", ["$injector", "$scope", "$rootScope", 
 				return $rootScope.usuarioSessao.usuarioEntradaUnica.perfilSelecionado.codigo;
 			}
 		},
+
+
+
+
+		{
+
+			titulo: 'Aguardando Autenticação',
+			icone: 'glyphicon glyphicon-pencil',
+			url: function() {
+				return '/';
+			},
+			countItens: true,
+			estaSelecionado: function () {
+
+				return $location.path() === '/caixa-entrada';
+			},
+			visivel: function(){
+
+				return [
+					app.utils.Perfis.PRESIDENTE,
+					].indexOf($rootScope.usuarioSessao.usuarioEntradaUnica.perfilSelecionado.codigo) > -1;
+			},
+			condicaoTramitacao: function() {
+
+				if ($rootScope.usuarioSessao.usuarioEntradaUnica.perfilSelecionado.codigo === app.utils.Perfis.PRESIDENTE)
+					return app.utils.CondicaoTramitacao.AGUARDANDO_ASSINATURA_PRESIDENTE;
+			},
+			deveFiltrarPorUsuario: true,
+			codigoPerfilSelecionado: function(){
+
+				return $rootScope.usuarioSessao.usuarioEntradaUnica.perfilSelecionado.codigo;
+			}
+		},
+
+
+
+
+
 		{
 
 			titulo: 'Em análise',
@@ -217,6 +256,10 @@ licenciamento.controller("AppController", ["$injector", "$scope", "$rootScope", 
 
 					return '/analise-diretor';
 
+				}else if ($rootScope.usuarioSessao.usuarioEntradaUnica.perfilSelecionado.codigo === app.utils.Perfis.PRESIDENTE){
+
+					return '/analise-presidente';
+
 				}
 			},
 			countItens: true,
@@ -225,14 +268,16 @@ licenciamento.controller("AppController", ["$injector", "$scope", "$rootScope", 
 				return $location.path().indexOf('/analise-geo') > -1 ||
 					$location.path().indexOf('/analise-tecnica') > -1 ||
 					$location.path().indexOf('/analise-gerente') > -1 ||
-					$location.path().indexOf('/analise-diretor') > -1;
+					$location.path().indexOf('/analise-diretor') > -1 ||
+					$location.path().indexOf('/analise-presidente') > -1;
 			},
 			visivel: function() {
 
 				return [app.utils.Perfis.ANALISTA_GEO,
 				app.utils.Perfis.ANALISTA_TECNICO,
 				app.utils.Perfis.GERENTE,
-				app.utils.Perfis.DIRETOR].indexOf($rootScope.usuarioSessao.usuarioEntradaUnica.perfilSelecionado.codigo) > -1;
+				app.utils.Perfis.DIRETOR,
+				app.utils.Perfis.PRESIDENTE].indexOf($rootScope.usuarioSessao.usuarioEntradaUnica.perfilSelecionado.codigo) > -1;
 			},
 			condicaoTramitacao: function () {
 
@@ -242,6 +287,8 @@ licenciamento.controller("AppController", ["$injector", "$scope", "$rootScope", 
 					return app.utils.CondicaoTramitacao.EM_ANALISE_TECNICA;
 				else if($rootScope.usuarioSessao.usuarioEntradaUnica.perfilSelecionado.codigo === app.utils.Perfis.GERENTE)
 					return app.utils.CondicaoTramitacao.EM_ANALISE_GERENTE;
+				else if($rootScope.usuarioSessao.usuarioEntradaUnica.perfilSelecionado.codigo === app.utils.Perfis.PRESIDENTE)
+					return app.utils.CondicaoTramitacao.EM_ANALISE_PRESIDENTE;
 			},
 			deveFiltrarPorUsuario: true,
 			codigoPerfilSelecionado: function(){
@@ -292,12 +339,12 @@ licenciamento.controller("AppController", ["$injector", "$scope", "$rootScope", 
 			},
 			visivel: function(){
 
-				return $rootScope.usuarioSessao.usuarioEntradaUnica.perfilSelecionado.codigo === app.utils.Perfis.APROVADOR && LICENCIAMENTO_CONFIG.usuarioSessao.autenticadoViaToken;
+				return $rootScope.usuarioSessao.usuarioEntradaUnica.perfilSelecionado.codigo === app.utils.Perfis.PRESIDENTE && LICENCIAMENTO_CONFIG.usuarioSessao.autenticadoViaToken;
 			},
 
 			condicaoTramitacao: function(){
 
-				return app.utils.CondicaoTramitacao.AGUARDANDO_ASSINATURA_APROVADOR;
+				return app.utils.CondicaoTramitacao.AGUARDANDO_ASSINATURA_PRESIDENTE;
 			},
             deveFiltrarPorUsuario: true
 		},